LPS-99426 Make it possible to click on overlapping overlays
This can happen if you have two targetable elements that are close
together; for example, say we have "target A" and "target B" close
together, and the overlay drawn on top of "target B" ends up partially
covering "target A":

                +---------------+----------------+
                |...............|     topper     |
    +-----------+----+..........+--+-------------+
    |                |.............|             |
    |                |.............|             |
    |    target A    |.............|  target B   |
    |                |.............|             |
    |                |.............|             |
    +-----------+----+-------------+-------------+
                |                                |
                |                                |
                |                                |
                |                                |
                |            popover             |
                |                                |
                |                                |
                |                                |
                |                                |
                |                                |
                +--------------------------------+

In this case we need to make sure that the entire overlay is transparent
to clicks -- except for the topper, the popover and the button itself --
so that they can go through to the other target. This transparent region
is the one marked with periods in the diagram above.